# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test ihold",
        "version": "1.0"
    },
    "paths": {
        "/api": {},
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "List all tasks",
                "operationId": "be68905d866e856db6c738172b8d929a",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a tasks",
                "operationId": "ebbaabc58e3db61374a57c53544959b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "com imagem"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "conteúdo com imagem"
                                    },
                                    "file": {
                                        "type": "file",
                                        "example": "image.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Tasks registered successfully"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get information about an task",
                "operationId": "32c40e2d1129b80c1dcabd538df63a33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Tasks retrieved successfully"
                            }
                        }
                    },
                    "404": {
                        "description": "Tasks not found"
                    }
                }
            }
        },
        "/api/tasks/{id}/": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update a tasks",
                "operationId": "f6c767ef6bb12caf46f5c7af0a84a035",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Novo título"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nova Descrição"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "CONCLUÍDO"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Task updated successfully"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a tasks",
                "operationId": "b237514cd6ad68271958bed33c43d89e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tasks to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Tasks deleted successfully"
                            }
                        }
                    },
                    "404": {
                        "description": "Tasks not found"
                    }
                }
            }
        }
    }
}